+++ sumo/src/utils/xml/XMLSubSys.cpp
@@ -46,7 +46,7 @@
 // static member variables
 // ===========================================================================
 std::vector<SUMOSAXReader*> XMLSubSys::myReaders;
-unsigned int XMLSubSys::myNextFreeReader;
+int XMLSubSys::myNextFreeReader;
 XERCES_CPP_NAMESPACE::SAX2XMLReader::ValSchemes XMLSubSys::myValidationScheme = XERCES_CPP_NAMESPACE::SAX2XMLReader::Val_Auto;
 XERCES_CPP_NAMESPACE::SAX2XMLReader::ValSchemes XMLSubSys::myNetValidationScheme = XERCES_CPP_NAMESPACE::SAX2XMLReader::Val_Auto;
 
@@ -115,7 +115,7 @@ XMLSubSys::runParser(GenericSAXHandler& handler,
                      const std::string& file, const bool isNet) {
     try {
         XERCES_CPP_NAMESPACE::SAX2XMLReader::ValSchemes validationScheme = isNet ? myNetValidationScheme : myValidationScheme;
-        if (myNextFreeReader == myReaders.size()) {
+        if (myNextFreeReader == (int)myReaders.size()) {
             myReaders.push_back(new SUMOSAXReader(handler, validationScheme));
         } else {
             myReaders[myNextFreeReader]->setValidation(validationScheme);
